Model/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
tqdm==4.65.0
pyyaml==6.0
opencv-python==4.8.1.78
albumentations==1.3.1
wandb==0.15.12
tensorboard==2.14.0
jupyter==1.0.0
ipykernel==6.24.0